Simplify and harmonize line based tools
